+++ packages/plugins/eslint-plugin-react-x/src/rules/no-nested-component-definitions/no-nested-component-definitions.ts
@@ -38,8 +38,8 @@ export function create(context: RuleContext<MessageID, []>) {
     | core.ComponentDetectionHint.DoNotIncludeJsxWithUndefinedValue
     | core.ComponentDetectionHint.RequireBothSidesOfLogicalExpressionToBeJsx
     | core.ComponentDetectionHint.RequireBothBranchesOfConditionalExpressionToBeJsx
-    | core.ComponentDetectionHint.DoNotIncludeFunctionDefinedInArrayPattern
-    | core.ComponentDetectionHint.DoNotIncludeFunctionDefinedInArrayExpression
+    | core.ComponentDetectionHint.DoNotIncludeFunctionDefinedAsArrayPatternElement
+    | core.ComponentDetectionHint.DoNotIncludeFunctionDefinedAsArrayExpressionElement
     | core.ComponentDetectionHint.DoNotIncludeFunctionDefinedAsArrayMapCallback;
 
   // Collectors to find all component definitions in the code